+++ grid2op/Exceptions/ambiguousActionExceptions.py
@@ -124,6 +124,14 @@ class InvalidCurtailment(AmbiguousAction):
 
     pass
 
+class InvalidBackendCallback(AmbiguousAction):
+    """
+    This is a more precise exception than :class:`AmbiguousAction` indicating that
+    the :class:`grid2op.BaseAction.BaseAction`
+    try to apply an invalid callback to the backend.
+    """
+    pass
+
 
 class GeneratorTurnedOnTooSoon(InvalidRedispatching):
     """