+++ grid2op/Space/GridObjects.py
@@ -2984,9 +2984,6 @@ class GridObjects:
             # the configuration equires to initialize the classes from the local environment path
             # this might be usefull when using pickle module or multiprocessing on Windows for example
             my_class = GridObjects._build_cls_from_import(name_res, gridobj._PATH_GRID_CLASSES)
-            # from grid2op.Observation import CompleteObservation
-            # if issubclass(CompleteObservation, cls):
-            #    print(f"DEBUG WINDOWS CI: init_grid {cls} gridobj._PATH_GRID_CLASSES is not None: {my_class}")
             if my_class is not None:
                 return my_class
         
@@ -3025,9 +3022,6 @@ class GridObjects:
         elif gridobj._PATH_GRID_CLASSES is not None:
             # If I end up it's because the environment is created with already initialized
             # classes.
-            # from grid2op.Observation import CompleteObservation
-            # if issubclass(CompleteObservation, cls):
-            #     print(f"DEBUG WINDOWS CI: init_grid {cls} gridobj._PATH_GRID_CLASSES is not None 2:  {my_class}")
             return cls._aux_init_grid_from_cls(gridobj, name_res)
         
         # legacy behaviour: build the class "on the fly"
@@ -3056,9 +3050,6 @@ class GridObjects:
         res_cls.process_detachment()
         res_cls.process_shunt_static_data()
         res_cls._check_convert_to_np_array()  # convert everything to numpy array
-        # from grid2op.Observation import CompleteObservation
-        # if issubclass(CompleteObservation, cls):
-        #     print(f"DEBUG WINDOWS CI: init_grid {cls} gridobj._PATH_GRID_CLASSES is not None 2:  {res_cls}")
         if force_module is not None:
             res_cls.__module__ = force_module  # hack because otherwise it says "abc" which is not the case
             # best would be to have a look at https://docs.python.org/3/library/types.html