+++ CHANGELOG.rst
@@ -102,6 +102,8 @@ Native multi agents support:
 
 [1.12.1] - 2025-xx-yy
 ----------------------
+- [ADDED] the possibility to act on the backend directly from the action
+  with the "act."
 - [IMPROVED] cleaner installation, relying only on "pyproject.toml"
 
 [1.12.0] - 2025-07-24